6251 | 14 files | M DumpSource2/schemas/client/CDOTA_Modifier_Luna_Eclipse.h,  M DumpSource2/schemas/client/CDOTA_Modifier_SandKing_SandStorm.h,  M DumpSource2/schemas/client/C_DOTA_Ability_Earthshaker_EchoSlam.h,  M DumpSource2/schemas/client/C_DOTA_Ability_Tusk_Snowball.h,  M DumpSource2/schemas/server/CDOTA_Ability_Earthshaker_EchoSlam.h,  M DumpSource2/schemas/server/CDOTA_Ability_Tusk_Snowball.h,  M DumpSource2/schemas/server/CDOTA_Modifier_Luna_Eclipse.h,  M DumpSource2/schemas/server/CDOTA_Modifier_SandKing_SandStorm.h,  M DumpSource2/schemas/server/CDOTA_Modifier_Snapfire_FiresnapCookie_ShortHop.h,  M game/bin/built_from_cl.txt,  M game/dota/bin/linuxsteamrt64/libclient_strings.txt,  M game/dota/bin/linuxsteamrt64/libserver_strings.txt,  M game/dota/pak01_dir.txt,  M game/dota/steam.inf
https://steamdb.info/patchnotes/16396251/

+++ DumpSource2/schemas/client/C_DOTA_Ability_Earthshaker_EchoSlam.h
@@ -1,6 +1,6 @@
 class C_DOTA_Ability_Earthshaker_EchoSlam : public C_DOTABaseAbility
 {
-	int32 echo_slam_damage_range;
+	float32 echo_slam_damage_range;
 	int32 echo_slam_echo_search_range;
 	int32 echo_slam_echo_range;
 	int32 echo_slam_echo_damage;
