+++ DumpSource2/schemas/server/CDOTA_Ability_Earthshaker_EchoSlam.h
@@ -1,6 +1,6 @@
 class CDOTA_Ability_Earthshaker_EchoSlam : public CDOTABaseAbility
 {
-	int32 echo_slam_damage_range;
+	float32 echo_slam_damage_range;
 	int32 echo_slam_echo_search_range;
 	int32 echo_slam_echo_range;
 	int32 echo_slam_echo_damage;
