+++ DumpSource2/schemas/server/CDOTA_Modifier_SandKing_SandStorm.h
@@ -1,6 +1,6 @@
 class CDOTA_Modifier_SandKing_SandStorm : public CDOTA_Buff
 {
-	int32 sand_storm_radius;
+	float32 sand_storm_radius;
 	int32 sand_storm_damage;
 	float32 damage_tick_rate;
 	GameTime_t m_flLastDamageTime;
